LayoutTests/inspector/file-system-project.html

<html>
<head>
<script src="../http/tests/inspector/inspector-test.js"></script>
<script src="../http/tests/inspector/debugger-test.js"></script>
<script src="../http/tests/inspector/workspace-test.js"></script>
<script>
function test()
{
    var MockIsolatedFileSystem = function(id, path, files)
    {
        this._id = id;
        this._path = path;
        this._files = files;
    };
    MockIsolatedFileSystem.prototype = {
        id: function()
        {
            return this._id;
        },

        path: function()
        {
            return this._path;
        },

        requestFileContent: function(path, callback)
        {
            callback(this._files[path]);
        },

        setFileContent: function(path, newContent, callback)
        {
            this._files[path] = newContent;
            callback();
        },

        requestFilesRecursive: function(path, callback)
        {
            callback(Object.keys(this._files));
        },
    }

    var MockIsolatedFileSystemManager = function() {};
    MockIsolatedFileSystemManager.prototype = {
        addMockFileSystem: function(path, files)
        {
            var id = this.fileSystemMapping.addFileSystemMapping(path);
            var fileSystem = new MockIsolatedFileSystem(id, path, files);
            this._fileSystems = this._fileSystems || {};
            this._fileSystems[path] = fileSystem;
            this.dispatchEventToListeners(WebInspector.IsolatedFileSystemManager.Events.FileSystemAdded, fileSystem);
        },

        removeMockFileSystem: function(path)
        {
            var fileSystem = this._fileSystems[path];
            delete this._fileSystems[path];
            this.dispatchEventToListeners(WebInspector.IsolatedFileSystemManager.Events.FileSystemRemoved, fileSystem);
        },

        __proto__: WebInspector.Object.prototype
    }

    function createIsolatedFileSystemManager()
    {
        InspectorTest.createWorkspace();
        var manager = new MockIsolatedFileSystemManager();
        manager.fileSystemMapping = new WebInspector.FileSystemMappingImpl();
        manager.fileSystemMapping._fileSystemMappingSetting = new InspectorTest.MockSetting({});
        InspectorTest.MockSetting
        manager.fileMapping = new WebInspector.FileMapping(manager.fileSystemMapping);
        manager.fileSystemWorkspaceProvider = new WebInspector.FileSystemWorkspaceProvider(manager, InspectorTest.testWorkspace, manager.fileMapping);
        return manager;
    }

    function dumpUISourceCodes(uiSourceCodes, next)
    {
        innerDumpUISourceCodes(uiSourceCodes, 0, next);

        function innerDumpUISourceCodes(uiSourceCodes, startIndex, next)
        {
            InspectorTest.addResult("");
            if (startIndex === uiSourceCodes.length) {
                next();
                return;
            }
            InspectorTest.dumpUISourceCode(uiSourceCodes[startIndex], innerDumpUISourceCodes.bind(this, uiSourceCodes, startIndex + 1, next));
        }
    }

    InspectorTest.runTestSuite([
        function testFileSystems(next)
        {
            var uiSourceCodes = [];
            var manager = createIsolatedFileSystemManager();
            var entry1 = new WebInspector.FileMapping.Entry("http://localhost/", "/var/www/localhost/");
            var entry2 = new WebInspector.FileMapping.Entry("http://www.example.com/", "/foo/bar/");
            manager.fileMapping.setMappingEntries([entry1, entry2]);
            InspectorTest.waitForWorkspaceUISourceCodeAddedEvent(uiSourceCodeAdded, 3);
            InspectorTest.addResult("Adding first file system.");
            manager.addMockFileSystem("/var/www", {"/localhost/foo.js": "<foo content>", "/bar.js": "<bar content>"});
            InspectorTest.addResult("Adding second file system.");
            manager.addMockFileSystem("/foo/bar", {"/baz.js": "<baz content>"});

            dumpUISourceCodes(uiSourceCodes, uiSourceCodesDumped);

            function uiSourceCodeAdded(uiSourceCode)
            {
                uiSourceCodes.push(uiSourceCode)
            }

            function uiSourceCodesDumped()
            {
                InspectorTest.addResult("UISourceCode uri to url mappings:");
                for (var i = 0; i < uiSourceCodes.length; ++i) {
                    var url = uiSourceCodes[i].url;
                    if (!url)
                        continue;
                    InspectorTest.addResult("    " + uiSourceCodes[i].uri() + " -> " + uiSourceCodes[i].url);
                }
                InspectorTest.addResult("UISourceCode url to uri mappings:");
                for (var i = 0; i < uiSourceCodes.length; ++i) {
                    var url = uiSourceCodes[i].url;
                    if (!url)
                        continue;
                    var uri = manager.fileMapping.uriForURL(url);
                    InspectorTest.addResult("    " + url + " -> " + uri);
                }
                var uiSourceCodesCount = InspectorTest.testWorkspace.uiSourceCodes().length;
                InspectorTest.addResult("Removing second file system.");
                manager.removeMockFileSystem("/var/www");
                InspectorTest.addResult("    number of uiSourceCodes in workspace after removing second file system: " + InspectorTest.testWorkspace.uiSourceCodes().length);
                InspectorTest.addResult("Removing first file system.");
                manager.removeMockFileSystem("/foo/bar");
                InspectorTest.addResult("    number of uiSourceCodes in workspace after removing first file system: " + InspectorTest.testWorkspace.uiSourceCodes().length);
                next();
            }
        },
    ]);
};
</script>
</head>
<body onload="runTest()">
<p>Tests file system project.</p>
</body>
</html>
